Adapt to platform partition as Ubuntu Core - part 2
the bootargs "rw earlycon" are needed for the kernel
6.8.0-1007-renesas #7-Ubuntu.

due to platform partition being #1, we need to shift the
mmc_seed_part and mmc_boot_part to the right by one.

add change so that the fdt file is loaded from
the platform partition.

Address to where load the fit was too close to the other
addresses, like fdt_addr_r, so we moved it to 0x61000000

bootm command has been changed to use the fdt file loaded
from the platform partition.

All this is thruplicated to G2L, G2LC and G2UL includes.